async/await prepared statement API (#390)
This patch adds a new `PreparedStatement` protocol to represent prepared statements and an `execute` function on `PostgresConnection` to prepare and execute statements.

To implement the features the patch also introduces a `PreparedStatementStateMachine` that keeps track of the state of a prepared statement at the connection level. This ensures that, for each connection, each statement is prepared once at time of first use and then subsequent uses are going to skip the preparation step and just execute it.

## Example usage

First define the struct to represent the prepared statement:

```swift
 struct ExamplePreparedStatement: PreparedStatement {
     static let sql = "SELECT pid, datname FROM pg_stat_activity WHERE state = $1"
     typealias Row = (Int, String)
     var state: String
     
    func makeBindings() -> PostgresBindings {
        var bindings = PostgresBindings()
        bindings.append(self.state)
        return bindings
    }

    func decodeRow(_ row: PostgresNIO.PostgresRow) throws -> Row {
        try row.decode(Row.self)
    }
}
```
then, assuming you already have a `PostgresConnection` you can execute it:
```swift
let preparedStatement = ExamplePreparedStatement(state: "active")
let results = try await connection.execute(preparedStatement, logger: logger)
for (pid, database) in results {
    print("PID: \(pid), database: \(database)")
}
```

---------

+++ Tests/PostgresNIOTests/New/Extensions/PSQLFrontendMessageDecoder.swift
@@ -142,7 +142,7 @@ extension PostgresFrontendMessage {
             }
 
             let parameters = (0..<parameterCount).map { _ -> ByteBuffer? in
-                let length = buffer.readInteger(as: UInt16.self)
+                let length = buffer.readInteger(as: UInt32.self)
                 switch length {
                 case .some(..<0):
                     return nil
